fix(lane_alignment): set user data for a character based on state WHILST character is received
User data should be true for data from converters (or sometimes test mode data).
Before, user data were set upon outputting data from the buffer,
but that led to cases where ILAS sequence was counted
as user_data, because it was saved without information about
state of the character and it was assumed the state is the same
upon saving and outputting the octet.